dom/tests/mochitest/chrome/test_moving_nodeList.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css"
                 type="text/css"?>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=684115
-->
<window title="Mozilla Bug 684115"
  xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">
  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <!-- test results are displayed in the html:body -->
  <body xmlns="http://www.w3.org/1999/xhtml">
  <a href="https://bugzilla.mozilla.org/show_bug.cgi?id=654370"
     target="_blank">Mozilla Bug 684115</a>
  </body>

  <!-- test code goes here -->
  <script type="application/javascript"><![CDATA[
      SimpleTest.waitForExplicitFinish();

      var firstWindow, secondWindow;
      function iframe_loaded() {
        if (!firstWindow || !secondWindow)
          return;
        var nodeList = firstWindow.document.childNodes;
        ok(!("expando" in nodeList), "shouldn't be able to see expandos on the NodeList");
        nodeList = firstWindow.wrappedJSObject.getNodeList();
        ok(("expando" in nodeList), "should be able to see expandos on the NodeList");
        is(nodeList.selectedIndex, -1, "can access selectedIndex in chrome");
        secondWindow.wrappedJSObject.tryToUseNodeList(nodeList, ok);
        nodeList = document.createElementNS("http://www.w3.org/1999/xhtml", "select").options;
        secondWindow.wrappedJSObject.tryToUseNodeList(nodeList, ok);
        SimpleTest.finish();
      }

  ]]></script>

  <iframe id="one" src="http://mochi.test:8888/tests/dom/tests/mochitest/general/file_moving_nodeList.html"
          onload="firstWindow = this.contentWindow; iframe_loaded()" />
  <iframe id="two" src="http://example.org/tests/dom/tests/mochitest/general/file_moving_nodeList.html"
          onload="secondWindow = this.contentWindow; iframe_loaded()" />
</window>
